refactor app diectory structure
