--- api/v1/app.py
@@ -1,3 +1,0 @@
-from api.v1.routes import *
-# app.run(host='127.0.0.1', port=1111)
-app.run()
