+++ api/v1/tests/test_bucketlists.py
@@ -1,7 +1,9 @@
 import json
+
 from flask_testing import TestCase
-from ...imports import app, envi, databases
-from ...v1.models import User
+
+from ..headers import app, envi, databases
+from ..models import User
 
 class TestBucketList(TestCase):
     def create_app(self):
