+++ manage.py
@@ -1,7 +1,7 @@
-from flask_script import Manager
 from flask_migrate import Migrate, MigrateCommand
+from flask_script import Manager
 
-from api.imports import app, databases
+from api.v1.headers import app, databases
 
 migrate = Migrate(app, databases)
 manager = Manager(app)
